+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/).
 
 ## Unreleased
+### Fixed
+- Optimizers
+    - _Braghin_
+        - Run the segmentator before creating the transformation.
+
 ## 1.13.2 - 2024-03-13
 ### Changed
 - Selectors